refactor: 용어 영속화 관련 클래스 패키지 변경
