+++ space-d/src/main/java/com/dnd/spaced/core/word/infrastructure/persistence/PopularWordRedisRepository.java
@@ -1,4 +1,4 @@
-package com.dnd.spaced.core.word.infrastructure;
+package com.dnd.spaced.core.word.infrastructure.persistence;
 
 import com.dnd.spaced.core.word.domain.repository.PopularWordRepository;
 import com.dnd.spaced.core.word.domain.repository.dto.PopularWordInfo;
