+++ space-d/src/main/java/com/dnd/spaced/core/word/infrastructure/persistence/PronunciationQuerydslRepository.java
@@ -1,4 +1,4 @@
-package com.dnd.spaced.core.word.infrastructure;
+package com.dnd.spaced.core.word.infrastructure.persistence;
 
 import static com.dnd.spaced.core.word.domain.QPronunciation.pronunciation;
