+++ space-d/src/main/java/com/dnd/spaced/core/word/infrastructure/persistence/WordExampleCrudRepository.java
@@ -1,4 +1,4 @@
-package com.dnd.spaced.core.word.infrastructure;
+package com.dnd.spaced.core.word.infrastructure.persistence;
 
 import com.dnd.spaced.core.word.domain.WordExample;
 import org.springframework.data.repository.CrudRepository;
