+++ space-d/src/main/java/com/dnd/spaced/core/word/infrastructure/persistence/WordExampleQuerydslRepository.java
@@ -1,4 +1,4 @@
-package com.dnd.spaced.core.word.infrastructure;
+package com.dnd.spaced.core.word.infrastructure.persistence;
 
 import static com.dnd.spaced.core.word.domain.QWordExample.wordExample;
