+++ space-d/src/main/java/com/dnd/spaced/core/word/infrastructure/persistence/WordRandomCrudRepository.java
@@ -1,4 +1,4 @@
-package com.dnd.spaced.core.word.infrastructure;
+package com.dnd.spaced.core.word.infrastructure.persistence;
 
 import com.dnd.spaced.core.word.domain.WordRandom;
 import org.springframework.data.repository.CrudRepository;
